+++ packages/bpk-component-accordion/index.js
@@ -1,5 +1,6 @@
 import BpkAccordion from './src/BpkAccordion';
 import BpkAccordionContainer from './src/BpkAccordionContainer';
 import BpkAccordionItem from './src/BpkAccordionItem';
+import withState from './src/withState';
 
-export { BpkAccordionContainer, BpkAccordion, BpkAccordionItem };
+export { BpkAccordionContainer, BpkAccordion, BpkAccordionItem, withState };