+++ packages/bpk-component-accordion/src/BpkAccordionContainer-test.js
@@ -19,7 +19,7 @@ describe('BpkAccordionContainer', () => {
     const tree = renderer.create(
       <BpkAccordionContainer>
         <div>Accordion Item 1</div>
-        <div expanded>Accordion Item 2</div>
+        <div initiallyExpanded>Accordion Item 2</div>
         <div>Accordion Item 3</div>
       </BpkAccordionContainer>,
     ).toJSON();
@@ -30,30 +30,8 @@ describe('BpkAccordionContainer', () => {
     const tree = renderer.create(
       <BpkAccordionContainer>
         <div>Accordion Item 1</div>
-        <div expanded>Accordion Item 2</div>
-        <div expanded>Accordion Item 3</div>
-      </BpkAccordionContainer>,
-    ).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('should render correctly with "allowMultiple" prop', () => {
-    const tree = renderer.create(
-      <BpkAccordionContainer allowMultiple>
-        <div>Accordion Item 1</div>
-        <div>Accordion Item 2</div>
-        <div>Accordion Item 3</div>
-      </BpkAccordionContainer>,
-    ).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('should render correctly with "allowMultiple" prop and custom initially expanded items', () => {
-    const tree = renderer.create(
-      <BpkAccordionContainer allowMultiple>
-        <div expanded>Accordion Item 1</div>
-        <div>Accordion Item 2</div>
-        <div expanded>Accordion Item 3</div>
+        <div initiallyExpanded>Accordion Item 2</div>
+        <div initiallyExpanded>Accordion Item 3</div>
       </BpkAccordionContainer>,
     ).toJSON();
     expect(tree).toMatchSnapshot();