+++ packages/bpk-component-accordion/src/BpkAccordionContainer.js
@@ -1,25 +1,50 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { Component, Children, cloneElement } from 'react';
 
-import BpkAccordionContainerSingle from './BpkAccordionContainerSingle';
-import BpkAccordionContainerMultiple from './BpkAccordionContainerMultiple';
+import BpkAccordion from './BpkAccordion';
 
-const BpkAccordionContainer = (props) => {
-  const { allowMultiple, ...rest } = props;
-
-  const Component = allowMultiple
-    ? BpkAccordionContainerMultiple
-    : BpkAccordionContainerSingle;
-
-  return <Component {...rest} />;
+const getInitiallyExpanded = (children) => {
+  const accordionItems = Children.toArray(children);
+  const result = accordionItems.reduceRight((prev, item) => (item.props.initiallyExpanded ? item : prev), {});
+  return (result || {}).key || null;
 };
+
+class BpkAccordionContainer extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      expanded: getInitiallyExpanded(this.props.children),
+    };
+
+    this.openAccordionItem = this.openAccordionItem.bind(this);
+    this.renderAccordionItem = this.renderAccordionItem.bind(this);
+  }
+
+  openAccordionItem(key) {
+    this.setState({ expanded: key });
+  }
+
+  renderAccordionItem(accordionItem) {
+    const expanded = this.state.expanded === accordionItem.key;
+    const onClick = () => this.openAccordionItem(accordionItem.key);
+
+    return cloneElement(accordionItem, { expanded, onClick });
+  }
+
+  render() {
+    const { children } = this.props;
+
+    return (
+      <BpkAccordion>
+        {Children.toArray(children).map(this.renderAccordionItem)}
+      </BpkAccordion>
+    );
+  }
+}
 
 BpkAccordionContainer.propTypes = {
-  allowMultiple: PropTypes.bool,
-};
-
-BpkAccordionContainer.defaultProps = {
-  allowMultiple: false,
+  children: PropTypes.node.isRequired,
 };
 
 export default BpkAccordionContainer;